implement constructor ref
All the neccessary things needed to implement the constructor reference were done. Changes to the IDatabase interface and ILine objects were required. There is a bit of an ugly fix for evaluating the references within the database objects themselves, this might better be placed in the execution of the moveCom object possibly.