add initial ObjectEntry objects and start writing subclasses
The nature of MongoDB required some sort of nested document model in the code. This manifested as the ObjectEntry class and it's subclasses. Eventually DataEntry classes will be able to convert themselves to objectEntry in order to insert into columns and to provide a model to get data from fields in mongodb and oher document based sources, eg. JSON.